Update paths to moved ProfileSettingsForm
