+++ src/containers/ProfileSettingsPage/ProfileSettingsForm/ProfileSettingsForm.module.css
@@ -1,4 +1,4 @@
-@import '../../styles/propertySets.css';
+@import '../../../styles/propertySets.css';
 
 .root {
   margin-top: 24px;
