+++ src/containers/ProfileSettingsPage/ProfileSettingsPage.js
@@ -8,7 +8,6 @@ import { propTypes } from '../../util/types';
 import { ensureCurrentUser } from '../../util/data';
 import { isScrollingDisabled } from '../../ducks/UI.duck';
 
-import { ProfileSettingsForm } from '../../forms';
 import {
   Page,
   UserNav,
@@ -20,6 +19,8 @@ import {
   NamedLink,
 } from '../../components';
 import TopbarContainer from '../../containers/TopbarContainer/TopbarContainer';
+
+import ProfileSettingsForm from './ProfileSettingsForm/ProfileSettingsForm';
 
 import { updateProfile, uploadImage } from './ProfileSettingsPage.duck';
 import css from './ProfileSettingsPage.module.css';
